feat: chunk refactor
