+++ src/core/modules/game_entites_stroll.js
@@ -8,7 +8,7 @@ import { context, current_three_character } from '../game/game.js'
 import {
   get_optional_terrain_height,
   get_terrain_height,
-} from '../utils/terrain/chunk_utils.js'
+} from '../utils/terrain/world_utils.js'
 
 const MOVE_INTERVAL = 5000 // 5 seconds in milliseconds
 const MOVE_PROBABILITY = 0.1 // 10% chance to move
